requirements.txt

bedrock-agentcore
strands-agents
boto3